+++ .idea/modules.xml
@@ -4,7 +4,6 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/YoutubeIntegration.iml" filepath="$PROJECT_DIR$/YoutubeIntegration.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/youtube-integration-in-Android.iml" filepath="$PROJECT_DIR$/youtube-integration-in-Android.iml" />
     </modules>
   </component>
