+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.techiesatish.youtubeintegration"
+        applicationId "com.siddharth.divinecare"
         minSdkVersion 18
         targetSdkVersion 28
         versionCode 1
