+++ app/src/main/AndroidManifest.xml
@@ -1,19 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.techiesatish.youtubeintegration">
+    package="com.siddharth.divinecare">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".AppController"
+        android:name="com.siddharth.divinecare.AppController"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity"></activity>
-        <activity android:name=".ChannelActivity">
+        <activity android:name="com.siddharth.divinecare.MainActivity"></activity>
+        <activity android:name="com.siddharth.divinecare.ChannelActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -22,7 +22,7 @@
 
 
         </activity>
-        <activity android:name=".VideoActivity">
+        <activity android:name="com.siddharth.divinecare.VideoActivity">
 
 
         </activity>
