+++ app/src/main/java/com/siddharth/divinecare/CustomListAdapter.java
@@ -1,8 +1,6 @@
-package com.techiesatish.youtubeintegration;
+package com.siddharth.divinecare;
 
-import android.support.v4.app.FragmentManager;
 import android.app.Activity;
-import android.app.FragmentTransaction;
 import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
