+++ app/src/main/java/com/siddharth/divinecare/VideoDetails.java
@@ -1,4 +1,4 @@
-package com.techiesatish.youtubeintegration;
+package com.siddharth.divinecare;
 
 /**
  * Created by Ansh on 29/10/2017 12:26 AM.
