+++ app/src/main/res/layout/activity_channel.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/youtube_fragment"
-    tools:context="com.techiesatish.youtubeintegration.ChannelActivity">
+    tools:context="com.siddharth.divinecare.ChannelActivity">
 
     <ListView
         android:layout_width="wrap_content"
