+++ app/src/main/res/layout/activity_video.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.techiesatish.youtubeintegration.VideoActivity">
+    tools:context="com.siddharth.divinecare.VideoActivity">
 
 
     <com.google.android.youtube.player.YouTubePlayerView
