+++ app/src/main/res/layout/fragment_youtube.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.techiesatish.youtubeintegration.YoutubeFragment">
+    tools:context="com.siddharth.divinecare.YoutubeFragment">
 
     <FrameLayout
         android:name="com.google.android.youtube.player.YouTubePlayerSupportFragment"
